catched errors in addOrAbortIfErrors
